requirements.txt
asgiref==3.11.1
Django==6.0.2
djangorestframework==3.16.1
joblib==1.5.3
numpy==2.4.2
scikit-learn==1.8.0
scipy==1.17.0
sqlparse==0.5.5
threadpoolctl==3.6.0
